Feature/clean graph extractor (#369)
* combine graph extractor interface and implementation

* use declared config parser

* drop graph extractor as integrator member
