+++ include/hydra/frontend/gvd_place_extractor.h
@@ -39,7 +39,7 @@
 
 #include "hydra/common/output_sink.h"
 #include "hydra/frontend/freespace_places_interface.h"
-#include "hydra/places/graph_extractor_config.h"
+#include "hydra/places/graph_extractor.h"
 #include "hydra/places/gvd_integrator_config.h"
 #include "hydra/places/gvd_voxel.h"
 #include "hydra/reconstruction/tsdf_interpolators.h"
@@ -50,7 +50,6 @@ namespace hydra {
 namespace places {
 // forward declare to avoid include
 class GvdIntegrator;
-class GraphExtractorInterface;
 }  // namespace places
 
 struct VolumetricWindow;
@@ -59,14 +58,14 @@ class GvdPlaceExtractor : public FreespacePlacesInterface {
  public:
   using PositionMatrix = Eigen::Matrix<double, 3, Eigen::Dynamic>;
   using Sink = OutputSink<uint64_t,
-                          const Eigen::Isometry3f&,
+                          const Eigen::Isometry3d&,
                           const places::GvdLayer&,
-                          const places::GraphExtractorInterface*>;
+                          const places::GraphExtractor&>;
 
   struct Config {
     std::string layer = DsgLayers::PLACES;
     places::GvdIntegratorConfig gvd;
-    config::VirtualConfig<places::GraphExtractorInterface> graph;
+    places::GraphExtractor::Config graph;
     config::VirtualConfig<TsdfInterpolator> tsdf_interpolator;
     size_t min_component_size = 3;
     bool filter_places = true;
@@ -102,7 +101,7 @@ class GvdPlaceExtractor : public FreespacePlacesInterface {
  protected:
   mutable std::mutex gvd_mutex_;
   places::GvdLayer::Ptr gvd_;
-  std::shared_ptr<places::GraphExtractorInterface> graph_extractor_;
+  places::GraphExtractor graph_extractor_;
   std::unique_ptr<places::GvdIntegrator> gvd_integrator_;
   std::unique_ptr<TsdfInterpolator> tsdf_interpolator_;
   std::unique_ptr<VolumetricWindow> map_window_;
